agents/05-memory-agent/local/requirements.txt
strands-agents
strands-agents-tools
boto3
bedrock-agentcore
